requirements.txt
python-binance==1.0.21
python-dotenv==1.0.0
pyyaml==6.0.1
aiosqlite==0.19.0
python-telegram-bot==20.7
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0